knowledge/ai-agents/workflows/theo/requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.0.14
python-dotenv>=1.0.0
pyyaml>=6.0.0
